Move project config routes
